+++ README.md
@@ -33,9 +33,9 @@
 ### DEMO
 
 <div style="display:flex" align="center">
-  <img src="https://github.com/LoakickBangu/madcamp_week_3/blob/master/images/1.gif" width="150">
+  <!-- <img src="https://github.com/LoakickBangu/madcamp_week_3/blob/master/images/1.gif" width="150">
   <img src="https://github.com/LoakickBangu/madcamp_week_3/blob/master/images/2.gif" width="150">
   <img src="https://github.com/LoakickBangu/madcamp_week_3/blob/master/images/3.gif" width="150">
   <img src="https://github.com/LoakickBangu/madcamp_week_3/blob/master/images/4.gif" width="150">
-  <img src="https://github.com/LoakickBangu/madcamp_week_3/blob/master/images/5.gif" width="150">
+  <img src="https://github.com/LoakickBangu/madcamp_week_3/blob/master/images/5.gif" width="150"> -->
 </div>
