+++ lib/base/base_divider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:week_3/helper.dart';
+import 'package:kaimarket/helper.dart';
 
 class BaseDivider extends StatelessWidget {
   final Color color;
