+++ lib/bloc/post_state.dart
@@ -1,8 +1,8 @@
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
 import 'package:adhara_socket_io/adhara_socket_io.dart';
-import 'package:week_3/utils/utils.dart';
-import 'package:week_3/models/post.dart';
+import 'package:kaimarket/utils/utils.dart';
+import 'package:kaimarket/models/post.dart';
 
 @immutable
 abstract class PostState extends Equatable {
