+++ lib/bloc/socket_state.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
 import 'package:adhara_socket_io/adhara_socket_io.dart';
-import 'package:week_3/utils/utils.dart';
+import 'package:kaimarket/utils/utils.dart';
 
 @immutable
 abstract class SocketState extends Equatable {
