+++ lib/chat/chat_card.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:week_3/chat/chat_view_page.dart';
-import 'package:week_3/styles/theme.dart';
-import 'package:week_3/utils/utils.dart';
-import 'package:week_3/models/chat.dart';
-import 'package:week_3/utils/base_height.dart';
+import 'package:kaimarket/chat/chat_view_page.dart';
+import 'package:kaimarket/styles/theme.dart';
+import 'package:kaimarket/utils/utils.dart';
+import 'package:kaimarket/models/chat.dart';
+import 'package:kaimarket/utils/base_height.dart';
 import 'package:intl/intl.dart';
 import 'package:cached_network_image/cached_network_image.dart';
