+++ lib/chat/chat_view_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:week_3/utils/base_height.dart';
-import 'package:week_3/utils/utils.dart';
+import 'package:kaimarket/utils/base_height.dart';
+import 'package:kaimarket/utils/utils.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:week_3/bloc/bloc.dart';
-import 'package:week_3/post/post_view_page.dart';
-import 'package:week_3/models/chat.dart';
+import 'package:kaimarket/bloc/bloc.dart';
+import 'package:kaimarket/post/post_view_page.dart';
+import 'package:kaimarket/models/chat.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:intl/intl.dart';
 
@@ -30,9 +30,6 @@ class _ChatViewPageState extends State<ChatViewPage> {
       EdgeInsets.only(left: 26, top: 12, bottom: 12, right: 26);
 
   final _partnerNameFont = TextStyle(fontSize: 20.0, color: Colors.grey[600]);
-  // final _chatFont = TextStyle(fontSize: 14.0, color: Colors.grey[500]);
-  // final _postFont = TextStyle(fontSize: 14.0, color: Colors.grey[600]);
-  // final _timeFont = TextStyle(fontSize: 10.0, color: Colors.grey[400]);
 
   List<Message> existMessages = [];
