+++ lib/layout/tab_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:week_3/utils/utils.dart';
-import 'package:week_3/styles/theme.dart';
+import 'package:kaimarket/utils/utils.dart';
+import 'package:kaimarket/styles/theme.dart';
 
 class TabButton extends StatelessWidget {
   final IconData icon;
