+++ lib/post/post_page.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:week_3/home/category_button.dart';
-import 'package:week_3/models/category.dart';
-import 'package:week_3/post/select_map_page.dart';
-import 'package:week_3/utils/base_height.dart';
-import 'package:week_3/post/photo_button.dart';
-import 'package:week_3/utils/utils.dart';
-import 'package:week_3/post/select_map_page.dart';
+import 'package:kaimarket/home/category_button.dart';
+import 'package:kaimarket/models/category.dart';
+import 'package:kaimarket/post/select_map_page.dart';
+import 'package:kaimarket/utils/base_height.dart';
+import 'package:kaimarket/post/photo_button.dart';
+import 'package:kaimarket/utils/utils.dart';
+import 'package:kaimarket/post/select_map_page.dart';
 import 'package:multi_image_picker/multi_image_picker.dart';
-import 'package:week_3/models/post.dart';
-import 'package:week_3/models/post.dart';
+import 'package:kaimarket/models/post.dart';
+import 'package:kaimarket/models/post.dart';
 import 'package:dio/dio.dart';
 
 class PostPage extends StatefulWidget {
