+++ lib/post/post_shimmer_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
-import 'package:week_3/utils/utils.dart';
+import 'package:kaimarket/utils/utils.dart';
 
 class PostShimmerCard extends StatelessWidget {
   @override
