+++ lib/post/post_view_page.dart
@@ -5,23 +5,23 @@ import 'package:carousel_pro/carousel_pro.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:shimmer/shimmer.dart';
-import 'package:week_3/bloc/post_event.dart';
-import 'package:week_3/models/chat.dart';
-import 'package:week_3/post/google_map_fixed.dart';
-import 'package:week_3/post/post_book_page.dart';
-import 'package:week_3/post/post_card.dart';
-import 'package:week_3/styles/theme.dart';
-import 'package:week_3/utils/utils.dart';
-import 'package:week_3/models/book.dart';
+import 'package:kaimarket/bloc/post_event.dart';
+import 'package:kaimarket/models/chat.dart';
+import 'package:kaimarket/post/google_map_fixed.dart';
+import 'package:kaimarket/post/post_book_page.dart';
+import 'package:kaimarket/post/post_card.dart';
+import 'package:kaimarket/styles/theme.dart';
+import 'package:kaimarket/utils/utils.dart';
+import 'package:kaimarket/models/book.dart';
 import 'dart:math' as math;
-import 'package:week_3/models/post.dart';
-import 'package:week_3/utils/dio.dart';
-import 'package:week_3/bloc/bloc.dart';
+import 'package:kaimarket/models/post.dart';
+import 'package:kaimarket/utils/dio.dart';
+import 'package:kaimarket/bloc/bloc.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:week_3/post/post_shimmer_card.dart';
-import 'package:week_3/chat/chat_view_page.dart';
-import 'package:week_3/post/post_page.dart';
+import 'package:kaimarket/post/post_shimmer_card.dart';
+import 'package:kaimarket/chat/chat_view_page.dart';
+import 'package:kaimarket/post/post_page.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 
 class PostViewPage extends StatefulWidget {
