+++ lib/splash.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:dio/dio.dart';
-import 'package:week_3/utils/utils.dart';
+import 'package:kaimarket/utils/utils.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'login/login_page.dart';
 import 'layout/default.dart';
