+++ lib/utils/dio.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:week_3/utils/utils.dart';
+import 'package:kaimarket/utils/utils.dart';
 
 //개발 모드에 따른 hostUrl
 const bool bDebug = true;
