+++ lib/wish/wish_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:week_3/utils/utils.dart';
-import 'package:week_3/models/post.dart';
-import 'package:week_3/post/post_view_page.dart';
-import 'package:week_3/post/post_card.dart';
-import 'package:week_3/bloc/user_bloc.dart';
+import 'package:kaimarket/utils/utils.dart';
+import 'package:kaimarket/models/post.dart';
+import 'package:kaimarket/post/post_view_page.dart';
+import 'package:kaimarket/post/post_card.dart';
+import 'package:kaimarket/bloc/user_bloc.dart';
 import 'package:bloc/bloc.dart';
-import 'package:week_3/bloc/bloc.dart';
+import 'package:kaimarket/bloc/bloc.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:fluttertoast/fluttertoast.dart';
