+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: week_3
+name: kaimarket
 description: A new Flutter project.
 
 # The following defines the version and build number for your application.
